backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.25.2
typing-extensions==4.8.0
newsapi-python==0.2.6
vaderSentiment==3.3.2
praw==7.7.1
beautifulsoup4==4.12.2
requests==2.31.0
pyportfolioopt==1.5.5
yfinance==0.2.18
scikit-learn==1.3.0
requests==2.31.0
